refactor(net): move `is_writable` function to `channel.rs` module
